+++ lib/quds_db/db_fields/json_field.dart
@@ -90,4 +90,6 @@ class JsonField extends _JsonField<Map> {
     result.parametersBuilder = () => [...getParameters(), key, value];
     return result;
   }
+
+  StringField asStringFieldForQuery() => StringField(columnName: columnName);
 }